bug 16500 finished: add action to import the sql file.
git-svn-id: http://talendforge.org/svn/top/trunk@52064 f6f1c999-d317-4740-80b0-e6d1abc6f99e

+++ org.talend.dataprofiler.core/src/org/talend/dataprofiler/core/ui/action/actions/handle/SimpleHandle.java
@@ -76,7 +76,7 @@ public class SimpleHandle implements IDuplicateHandle, IDeletionHandle {
      * @see org.talend.dataprofiler.core.ui.action.actions.handle.IDeletionHandle#delete()
      */
     public boolean delete() throws Exception {
-        if (file.exists() && isPhysicalDelete()) {
+        if (file.exists() && isPhysicalDelete() && !file.isReadOnly()) {
             file.delete(true, null);
         }
